Unify macfs node types and simplify DMG plan/build
hfsplus and apfs each defined their own Source/Bytes/Node types,
and pkg/dmg kept a fourth near-identical copy plus a recursive
converter and per-filesystem ID write-back map just to bridge
between them. planImage duplicated the same pipeline twice, once
per filesystem, differing only by package qualifier.

Add pkg/macfs.go with the shared Source, Bytes, FromFile, File,
and Node types (ID widened to uint64, plus Child helper); hfsplus
and apfs now alias them instead of redeclaring. This removes
planImage's imageNode converter and the two copy-pasted build
pipelines in pkg/dmg/filesystem.go. HFS+ now range-checks its own
catalog numbering instead of the caller doing it.

Also collapse dmg.go's three node-construction sites into one
recursive nodeFromInfo, call sourceMetadata exactly once and skip
it for symlinks, and export dmg.ParseFileSystem so CreateDMG and
the CLI flag validator share one filesystem-name parser.

+++ pkg/macfs/hfsplus/hfsplus_test.go
@@ -204,11 +204,11 @@ func TestAssignIDsIsStableAndIndependentOfContents(t *testing.T) {
 			{Name: "a", Data: Bytes([]byte(contents))},
 		}}}
 	}
-	ids := func(v Volume) map[string]uint32 {
+	ids := func(v Volume) map[string]uint64 {
 		if err := AssignIDs(&v); err != nil {
 			t.Fatal(err)
 		}
-		out := map[string]uint32{}
+		out := map[string]uint64{}
 		var walk func(string, *Node)
 		walk = func(prefix string, n *Node) {
 			for _, c := range n.Children {
